august21-event/project/assets/skybox/requirements.txt
numpy == 2.0.0
pillow == 10.4.0
